Adding Order Module Migration

+++ Cooklee.Data/Entities/Order/Order.cs
@@ -13,7 +13,7 @@ namespace Cooklee.Data.Entities.Order
             //
         }
 
-        public Order(string clientEmail, OrderAddress shippingAddress, ICollection<OrderItem> items, decimal subTotal)
+        public Order(string clientEmail, OrderAddress shippingAddress, ICollection<OrderItem> items, float subTotal)
         {
             ClientEmail = clientEmail;
             ShippingAddress = shippingAddress;
@@ -26,9 +26,9 @@ namespace Cooklee.Data.Entities.Order
         public OrderStatus Status { get; set; } = OrderStatus.Pending;
         public OrderAddress ShippingAddress { get; set; }
         public ICollection<OrderItem> Items { get; set; } = new HashSet<OrderItem>();
-        public decimal SubTotal { get; set; }
+        public float SubTotal { get; set; }
         //public int? DeliveryMethodId { get; set; }
         public DeliveryMethod DeliveryMethod { get; set; }
-        public decimal GetTotal() => SubTotal + DeliveryMethod.Cost;
+        public float GetTotal() => SubTotal + DeliveryMethod.Cost;
     }
 }
